// src/interfaces/Order/OrderRequest.ts
export interface OrderRequest {
  userId: string;
  userName: string;
  firstName: string;
  lastName: string;
  bikeId: string;
  bikeName: string;
  bikeMake: string;
  bikeModel: string;
  serialNumber: string;
  yearOfProduction: number;
  labelYearOfProduction: string;
  note: string;
  formChange: string;
}